+++ src/commands/pwd.ts
@@ -1,12 +1,7 @@
-import { execSync } from 'node:child_process';
-
-/** Returns the root of the repo or worktree */
-function getGitDirRoot(): string {
-	return execSync('git rev-parse --show-toplevel', { encoding: 'utf-8' }).trim();
-}
+import { getGitDirRoot } from '../utils/git';
 
 /** Handler for pwd */
-export function pwd() {
-	const pwd = getGitDirRoot();
-	console.log(pwd);
+export async function pwd() {
+  const pwd = await getGitDirRoot();
+  console.log(pwd);
 }